+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Don't make max line length depend on line break format
 * Match SendGrid transaction IDs
+* `idnSupported()` now static, as previously documented
 
 ## Version 6.0.1 (September 14th 2017)
 * Use shorter Message-ID headers (with more entropy) to avoid iCloud blackhole bug